refactor(trainer): move Ray Data I/O configs to shared ray_data_io.py
Address Hong's review feedback (PR #1415): extract ParquetReadConfig,
BatchIterConfig, and DataloadingConfig from tabular_trainer.py into a
new shared schema module so other workflow tasks (llm_trainer, etc.) can
reuse them without importing from a task-specific file.

- Add workflow/schema/ray_data_io.py with ParquetReadConfig, BatchIterConfig,
  DataloadingConfig — matches internal sdk/workflow/schema/v2alpha1/ray_data_io.py
  pattern
- Add workflow/schema/tests/ray_data_io_test.py with 6 tests for the
  extracted classes
- tabular_trainer.py imports the three classes from ray_data_io; inline
  definitions removed; __all__ re-exports them for backwards compatibility
- tabular_trainer_test.py imports the three classes from ray_data_io

49 tests total, all passing.
